Rearranged DashboardCard components and adjusted the visualization to new card content
